requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
Django==3.2.8
djangorestframework==3.12.4
environs==9.3.4
idna==3.3
marshmallow==3.13.0
multidict==5.2.0
Pillow==8.4.0
pycodestyle==2.8.0
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
